+++ lib/src/widgets/toolbar/toggle_style_button.dart
@@ -148,5 +148,6 @@ Widget defaultToggleStyleButtonBuilder(
     icon: Icon(icon, size: iconSize, color: iconColor),
     fillColor: fill,
     onPressed: onPressed,
+    borderRadius: iconTheme?.borderRadius ?? 2,
   );
 }